zadanie1/requirements.txt
Django==2.2.1
django-rest-framework==0.1.0
djangorestframework==3.9.3
pip==19.0.3
setuptools==2019.1
sqlparse==0.3.0